Module Backend.Repo

A repo abstraction.

type t = repo

Repo opening and closing functions

val v : Irmin__Conf.t -> t Lwt.t

v config is a function returning fresh store handles, with the configuration config, which is provided by the backend.

val close : t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

Getters from repo to backend store in ro mode

val contents_t : t -> Irmin.Perms.read Contents.t

val node_t : t -> Irmin.Perms.read Node.t

val commit_t : t -> Irmin.Perms.read Commit.t

val config : t -> Irmin__Conf.t

val batch : t -> ( Irmin.Perms.read_write Contents.t -> Irmin.Perms.read_write Node.t -> Irmin.Perms.read_write Commit.t -> 'a Lwt.t ) -> 'a Lwt.t

A getter from repo to backend stores in rw mode.

val branch_t : t -> Branch.t

A branch store getter from repo