Module Private.Commit

Private commit store.

Commit Store

include Irmin__.S.CONTENT_ADDRESSABLE_STORE

type -'a t

The type for content-addressable backend stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

val add : [> Irmin__.Import.write ] t -> value -> key Lwt.t

Write the contents of a value to the store. It's the responsibility of the content-addressable store to generate a consistent key.

val unsafe_add : [> Irmin__.Import.write ] t -> key -> value -> unit Lwt.t

Same as add but allows to specify the key directly. The backend might choose to discared that key and/or can be corrupt if the key scheme is not consistent.

include Irmin__.S.Store_properties.CLEARABLE with type 'a t := 'a t

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.

val merge : [> Irmin__.Import.read_write ] t -> info:Irmin.Info.f -> key option Irmin.Merge.t

merge is the 3-way merge function for commit keys.

module Key : Irmin.Hash.TYPED with type t = key and type value = value

Key provides base functions for commit keys.

module Val : Irmin__.Commit_intf.S with type t = value and type hash = key

Val provides functions for commit values.

module Node : Irmin__.Node.STORE with type key = Val.hash

Node is the underlying node store.