Module Maker.RO

type ctx

type -'a t = {

uri : Uri.t;

item : string;

items : string;

ctx : ctx option;

}

include Irmin.Read_only.S with type 'a t := 'a t

type -'a t

The type for stores. The 'a phantom type carries information about the store mutability.

type key

The type for keys.

type value

The type for raw values.

val mem : [> Irmin__.Import.read ] t -> key -> bool Lwt.t

mem t k is true iff k is present in t.

val find : [> Irmin__.Import.read ] t -> key -> value option Lwt.t

find t k is Some v if k is associated to v in t and None is k is not present in t.

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

module HTTP : Helper with type ctx = ctx

val uri : 'a t -> Uri.t

val item : 'a t -> string

val items : 'a t -> string

val key_str : key -> string

val val_of_str : value T.of_string

val v : ?⁠ctx:ctx -> Uri.t -> string -> string -> 'a t Lwt.t