Module type S.APPEND_ONLY_STORE_MAKER

Parameters

Client : Cohttp_lwt.S.Client

K : Irmin.Hash.S

V : Irmin.Type.S

Signature

type -'a t

type key = K.t

type value = V.t

type ctx = Client.ctx

val mem : [> Irmin_http__.Import.read ] t -> key -> bool Lwt.t

val find : [> Irmin_http__.Import.read ] t -> key -> value option Lwt.t

val add : 'a t -> value -> key Lwt.t

val unsafe_add : 'a t -> key -> value -> unit Lwt.t

val v : ?⁠ctx:ctx -> Uri.t -> string -> string -> 'a t Lwt.t

type 'a t

val close : 'a t -> unit Lwt.t

close t frees up all the resources associated with t. Any operations run on a closed handle will raise Closed.

type 'a t

val batch : Irmin__.Import.read t -> ([ Irmin__.Import.read | Irmin__.Import.write ] t -> 'a Lwt.t) -> 'a Lwt.t

batch t f applies the writes in f in a separate batch. The exact guarantees depend on the implementation.

type 'a t

val clear : 'a t -> unit Lwt.t

Clear the store. This operation is expected to be slow.