Module Irmin_http__Irmin_http_server.Make

Create an HTTP server, serving the contents of an Irmin database.

Parameters

HTTP : Cohttp_lwt.S.Server

S : Irmin.S

Signature

type repo = S.Repo.t

The type for Irmin repository.

type t = HTTP.t

The type for HTTP configuration.

val v : ?⁠strict:bool -> repo -> t

v repo returns the configuration for a server serving the contents of repo. If strict is set, incoming connections will fail if they do not have the right X-IrminVersion headers.