Module Irmin_http.Server

Parameters

module HTTP : Cohttp_lwt.S.Server

module S : Irmin.S

Signature

type repo = S.Repo.t

The type for Irmin repository.

type t = HTTP.t

The type for HTTP configuration.

val v : ?⁠strict:bool -> repo -> t

v repo returns the configuration for a server serving the contents of repo. If strict is set, incoming connections will fail if they do not have the right X-IrminVersion headers.