Module type Irmin_http.SERVER

type repo

The type for Irmin repository.

type t

The type for HTTP configuration.

val v : ?strict:bool -> repo -> t

v repo returns the configuration for a server serving the contents of repo. If strict is set, incoming connections will fail if they do not have the right X-IrminVersion headers.